Fix max weight limit type

+++ src/util/precompile.js
@@ -680,9 +680,9 @@ export const ISubnetABI = [
     name: "getMaxWeightLimit",
     outputs: [
       {
-        internalType: "uint64",
+        internalType: "uint16",
         name: "",
-        type: "uint64",
+        type: "uint16",
       },
     ],
     stateMutability: "view",
@@ -1126,9 +1126,9 @@ export const ISubnetABI = [
         type: "uint16",
       },
       {
-        internalType: "uint64",
+        internalType: "uint16",
         name: "maxWeightLimit",
-        type: "uint64",
+        type: "uint16",
       },
     ],
     name: "setMaxWeightLimit",
